Fix writes if namespace is too small
The namespace size detection code returned a generic ZstorError, which
was later not picked up when removing the database from the list of
heatlhy ones. This caused the write to proceed, even though it
shouldn't. Ultimately this leads to missing shards in the metadata,
which then leads to retrieval failure in #29

Signed-off-by: Lee Smet <lee.smet@hotmail.com>