Remove bad debug assert
The debug assert fails now that Key is an enum of both V1 and V2 key
types. It fails because it is a static size assert, so V1 zdb always
triggers it, and furthermore the enum discrimanent takes up a byte,
which means the `Key` type actually occupies 12 bytes (3 bytes of
padding due to the 4 byte allignment), meaning debug builds _always_
crash here.

Signed-off-by: Lee Smet <lee.smet@hotmail.com>